Books now once again work correctly in the CC and EC. Travesty now uses the proper weapon specials. Specials when in natual form have also been toned down Controlled creatures no longer lose focus when attacking on kill/attack control order Temporarily removed looting rights for healing as this is suspected in causing a shard lockup. It will either be refactored or investigated further Fixed issue where boat courses were not working properly and the tillerman was piloting the boat in the wrong DirectionL Boat course no longer works in the corgul region Corgul Island map now indicates when the encounter will Expire Simplified code used to determine if an item or subclass of an item can be crafted

+++ Scripts/Services/HuntmasterChallenge/ResourceSatchels.cs
@@ -32,7 +32,7 @@ namespace Server.Items
             int total = base.GetTotal(type);
 
             if (type == TotalType.Weight)
-                total -= total * (int)(_WeightReduction / 100.0);
+                total -= total * _WeightReduction / 100;
 
             return total;
         }
